phase-1: restore CI gate integrity with pinned toolchain
Pin ruff/mypy/pytest/pytest-cov in a dev extra (pyproject.toml), point CI at
".[webapp,dev]", and clear the 766 ruff violations surfaced by ruff 0.16.1's
expanded default rule set. 671 auto-fixed, 101 redundant noqa: E402 comments
removed, 72 manual violations triaged by category (BLE001/S110 narrowed per
site; PySAM guards kept broad with documented noqa since PySAM raises bare
Exception). Suite count byte-identical: 634 passed, 18 deselected, 3 xfailed.
CI had run on a red pipeline since 2026-07-26 because the lint gate was
unpinned and ruff expanded its defaults between releases.

+++ scripts/python/integration/analyze_ninhsim_cppa.py
@@ -14,7 +14,6 @@ import argparse
 import json
 from pathlib import Path
 
-
 REPO_ROOT = Path(__file__).resolve().parent.parent.parent.parent
 DEFAULT_STRIKE_ADJUSTMENT_FRACTIONS = [-0.15, -0.10, -0.05, 0.0, 0.05]
 DEFAULT_ANNUAL_GENERATION_DEGRADATION_FRACTION = 0.005
@@ -49,7 +48,7 @@ def _band_label(relative_to_ceiling_fraction: float) -> str:
         return "ceiling"
 
     percentage = abs(relative_to_ceiling_fraction) * 100.0
-    whole_percentage = int(round(percentage))
+    whole_percentage = round(percentage)
     if whole_percentage == percentage:
         percentage_text = str(whole_percentage)
     else:
